configuracao inicial para versionar a api
