+++ config/routes.rb
@@ -2,12 +2,13 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 
-  post 'auth/login', to: 'authentication#authenticate'
-
-  post 'signup', to: 'users#create'
-
-  resources :todos do
-    resources :items
+  # namespace the controllers without affecting the URI
+  scope module: :v1, constraints: ApiVersion.new('v1', true) do
+    resources :todos do
+      resources :items
+    end
   end
 
+  post 'auth/login', to: 'authentication#authenticate'
+  post 'signup', to: 'users#create'
 end
